Add DocUploadBundleTests
