+++ Package.swift
@@ -42,6 +42,7 @@ let package = Package(
             .product(name: "Zip", package: "Zip"),
             .product(name: "Dependencies", package: "swift-dependencies")
         ]),
-        .testTarget(name: "DocUploaderTests", dependencies: ["DocUploader", "DocUploadBundle"]),
+        .testTarget(name: "DocUploadBundleTests", dependencies: ["DocUploadBundle"]),
+        .testTarget(name: "DocUploaderTests", dependencies: ["DocUploader"]),
     ]
 )
